+++ wren-modeling-rs/core/src/logical_plan/utils.rs
@@ -11,8 +11,9 @@ use petgraph::Graph;
 use crate::mdl::lineage::DatasetLink;
 use crate::mdl::{
     manifest::{Column, Model},
-    Dataset, WrenMDL,
+    WrenMDL,
 };
+use crate::mdl::dataset::Dataset;
 
 pub fn map_data_type(data_type: &str) -> Result<DataType> {
     let result = match data_type {
