+++ wren-modeling-rs/core/src/mdl/manifest.rs
@@ -39,33 +39,6 @@ pub struct Model {
     pub properties: Vec<(String, String)>,
 }
 
-impl Model {
-    /// Physical columns are columns that can be selected from the model.
-    /// e.g. columns that are not a relationship column
-    pub fn get_physical_columns(&self) -> Vec<Arc<Column>> {
-        self.columns
-            .iter()
-            .filter(|c| c.relationship.is_none())
-            .map(Arc::clone)
-            .collect()
-    }
-
-    pub fn name(&self) -> &str {
-        &self.name
-    }
-
-    pub fn get_column(&self, column_name: &str) -> Option<Arc<Column>> {
-        self.columns
-            .iter()
-            .find(|c| c.name == column_name)
-            .map(Arc::clone)
-    }
-
-    pub fn primary_key(&self) -> Option<&str> {
-        self.primary_key.as_deref()
-    }
-}
-
 mod table_reference {
     use serde::{self, Deserialize, Deserializer, Serialize, Serializer};
